requirements.txt
# API Framework
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# NLP for PII Detection
spacy==3.7.2

# Database
pymongo==4.6.1

# Utilities
python-multipart==0.0.6
werkzeug==3.0.1

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0

# Security
cryptography==41.0.7